requirements-prod.txt
# Phase 10: Production Requirements
# Updated dependencies for production deployment

# ============================================
# Core Web Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ============================================
# Streamlit (existing)
# ============================================
streamlit==1.28.0
streamlit-authenticator==0.2.3

# ============================================
# HTTP & API
# ============================================
requests==2.31.0
httpx==0.25.0
aiohttp==3.9.0

# ============================================
# Data Processing
# ============================================
pandas==2.1.3
numpy==1.26.2

# ============================================
# Machine Learning
# ============================================
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Optional: Deep Learning (comment out if not needed)
# tensorflow==2.15.0
# torch==2.1.0

# ============================================
# ML Optimization & Explainability
# ============================================
optuna==3.4.0
shap==0.43.0
# lime==0.2.0.1  # Optional

# ============================================
# Visualization
# ============================================
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# ============================================
# Database
# ============================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# ============================================
# Caching & Session
# ============================================
redis==5.0.1
python-redis-cache==3.0.1

# ============================================
# Configuration & Security
# ============================================
python-dotenv==1.0.0
PyYAML==6.0.1
pyjwt==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# ============================================
# Monitoring & Logging
# ============================================
sentry-sdk[fastapi]==1.38.0
python-json-logger==2.0.7
loguru==0.7.2

# ============================================
# Performance
# ============================================
gunicorn==21.2.0
gevent==23.9.1

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
toml==0.10.2
click==8.1.7

# ============================================
# Production Tools
# ============================================
flower==2.0.1  # Celery monitoring (if using)
prometheus-client==0.19.0  # Metrics
